Add missing dependencies back into `package.json`
Add dependencies back in so that they're installed as expected. Also updated dependencies to be the latest versions of each. `react-native-swiper` was previously pointing at https://github.com/FuYaoDe/react-native-swiper which may or may not have been necessary, but which has now fallen behind in development over the original, maintained repository.